+++ mitra_cli/src/main.rs
@@ -56,6 +56,7 @@ async fn main() {
         SubCommand::DeleteOrphanedFiles(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::DeleteEmptyProfiles(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::PruneRemoteEmojis(cmd) => cmd.execute(&config, db_client).await.unwrap(),
+        SubCommand::ListLocalFiles(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::ListUnreachableActors(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::AddEmoji(cmd) => cmd.execute(&config, db_client).await.unwrap(),
         SubCommand::ImportEmoji(cmd) => cmd.execute(&config, db_client).await.unwrap(),