+++ mitra_models/src/media.rs
@@ -99,3 +99,39 @@ async fn find_orphaned_ipfs_objects(
         .collect::<Result<_, _>>()?;
     Ok(orphaned_ipfs_objects)
 }
+
+pub async fn get_local_files(
+    db_client: &impl DatabaseClient,
+) -> Result<Vec<String>, DatabaseError> {
+    let rows = db_client.query(
+        "
+        SELECT
+            unnest(array_remove(
+                ARRAY[
+                    avatar ->> 'file_name',
+                    banner ->> 'file_name'
+                ],
+                NULL
+            )) AS file_name
+        FROM actor_profile
+        WHERE
+            user_id IS NOT NULL
+            OR portable_user_id IS NOT NULL
+        UNION
+        SELECT file_name FROM media_attachment
+        JOIN actor_profile ON (media_attachment.owner_id = actor_profile.id)
+        WHERE
+            actor_profile.user_id IS NOT NULL
+            OR actor_profile.portable_user_id IS NOT NULL
+        UNION
+        SELECT image ->> 'file_name' AS file_name
+        FROM emoji
+        WHERE hostname IS NULL
+        ",
+        &[],
+    ).await?;
+    let filenames = rows.iter()
+        .map(|row| row.try_get("file_name"))
+        .collect::<Result<_, _>>()?;
+    Ok(filenames)
+}