+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `list-local-files` command.
+
 ### Fixed
 
 - Don't drop `Announce(Dislike)` activities.